% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_table.R
\name{anova_table}
\alias{anova_table}
\title{Generate an ANOVA Table}
\usage{
anova_table(models = list(), ..., output = c("raw", "plain", "fancy"))
}
\arguments{
\item{models}{A list of models to compare, in order from smallest to largest if nested. A named list can be used to override the default model names displayed in the table.}

\item{...}{Arguments to pass to `anova`}

\item{output}{Output format, either "raw", "plain", or "fancy". Defaults to "plain"}
}
\value{
A kable object for `output = "plain", "fancy"`, A tibble for `output = "raw"`
}
\description{
Creates a formatted table listing ANOVA statistics for the comparison of multiple models
}
