% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_var_names.R
\name{append_var_names}
\alias{append_var_names}
\title{Recode Value(s) as Missing}
\usage{
append_var_names(data_set, ..., prefix = "", suffix = "")
}
\arguments{
\item{data_set}{A tibble/data frame}
}
\value{
A tibble/data frame
}
\description{
Appends a string to all selected variable names
}
