% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_to_other.R
\name{collapse_to_other}
\alias{collapse_to_other}
\title{Collapse to Other}
\usage{
collapse_to_other(vect, n_categories, keep_numeric = TRUE)
}
\arguments{
\item{vect}{a vector}

\item{n_categories}{the total number of categories the returned vector will have}

\item{keep_numeric}{if keep_numeric is TRUE and the vector is numeric, the other category will be specified as a numeric value of repeated 8's that's greater than the maximum value instead of a character value}
}
\value{
A vector
}
\description{
Collapases a vector into the n-1 most frequent categories with all other categories becoming "other"
}
