% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab_percent.R
\name{crosstab_percent}
\alias{crosstab_percent}
\title{Calculate Percents for Contingency Tables}
\usage{
crosstab_percent(data_crosstab, n_var = "n", vars = c(), formatted = FALSE)
}
\arguments{
\item{data_crosstab}{A contingency table or tibble/data frame of a contingency table}

\item{n_var}{The name of the variable which contains frequencies}

\item{vars}{A character vector of variable names}

\item{formatted}{If "TRUE", percents of frequencies will be a character vector with rounded percents and a percent sign (\%), otherwise the raw decimal integers will be returned}
}
\value{
A contingency table as a tibble
}
\description{
Calculates percents for contingency tables (cross tabulations), by default selects the first n-2 variables (variable n is the number of observations, n-1 is the default omitted column)
}
