% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.R
\name{format_number}
\alias{format_number}
\title{Format Number}
\usage{
format_number(x, round_digits = c(3, 2, 1, 0), large_num_sep = ",")
}
\arguments{
\item{x}{A numeric vector}

\item{round_digits}{The number of digits to round numbers at each cutoff (< 1, < 10, < 100, >= 100). Default value is 3, 2, 1, 0.}

\item{large_num_sep}{A character to separate every three digits in large numbers. Default value is the American standard of using a comma (",").}
}
\value{
A character vector
}
\description{
Formats numeric vectors, rounding values to a reasonable amount of digits, and inserting separators for large numbers. Useful for inserting inline numeric values in R Markdown.
}
\examples{
format_number(pi * c(0.1, 1, 10, 100, 1000))
"0.314" "3.14"  "31.4"  "314"   "3,142"
}
