% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos_neg_format.R
\name{pos_neg_format}
\alias{pos_neg_format}
\title{Format Magnitude of Values}
\usage{
pos_neg_format(x, values = c("black", "grey35", "black"))
}
\arguments{
\item{x}{A numeric vector}

\item{values}{A character vector of colors to use for each interval}
}
\value{
A vector
}
\description{
An internal function for generating colors associated with how close a value is to the absolute maximum/minimum value in the vector
}
