% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_table.R
\name{regression_table}
\alias{regression_table}
\title{Generate a Regression Table}
\usage{
regression_table(
  reg_obj,
  table_caption = NULL,
  output = c("raw", "plain", "fancy"),
  fancy = FALSE
)
}
\arguments{
\item{reg_obj}{An object for a regreession type model}

\item{table_caption}{A caption to replace the default}

\item{output}{Output format, either "raw", "plain", or "fancy". Defaults to "plain"}

\item{fancy}{Deprecated - use `output`}
}
\value{
A kable object for `output = "plain", "fancy"`, A tibble for `output = "raw"`
}
\description{
Creates a formatted table listing variables in a regression type model, with their estimated coefficients and statistics
}
