% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_format.R
\name{sign_format}
\alias{sign_format}
\title{Format Significance Values}
\usage{
sign_format(
  x,
  breaks = c(1, 0.1, 0.05, 0.01, 0.001, 0),
  values = c("#808080", "#ffa500", "#9acd32", "#008000", "#006400"),
  na_value = "#808080"
)
}
\arguments{
\item{x}{A numeric vector with significance values}

\item{breaks}{A numeric vector with values at which each color break occurs}

\item{values}{A character vector of colors to use for each interval}

\item{na_value}{A character vector of the color to use for NA values}
}
\value{
A vector
}
\description{
An internal function for generating colors associated with significance values in hypothesis testing.
}
