% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_var_names.R
\name{std_var_names}
\alias{std_var_names}
\title{Standardize Variable Names}
\usage{
std_var_names(data_set)
}
\arguments{
\item{data_set}{A tibble/data frame}
}
\value{
A tibble/data frame
}
\description{
Formats variable names, replacing all spaces with underscores, and making all letters lowercase
}
