% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_table.R
\name{regression_table}
\alias{regression_table}
\title{Generate a Regression Table}
\usage{
regression_table(
  reg_obj,
  table_caption = NULL,
  fancy = FALSE,
  doc_format = "default"
)
}
\arguments{
\item{reg_obj}{An object for a regreession type model}

\item{table_caption}{A caption to replace the default}

\item{fancy}{Use kableExtra formatting to add colors for coeffiecents and p-values. Default value is `FALSE`.}

\item{doc_format}{Output format for R Markdown, refer to kableExtra format for more info, matches `knitr.table.format` by default}
}
\value{
A kable object
}
\description{
Creates a formatted table listing variables in a regression type model, with their estimated coefficients and statistics
}
